requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto3==1.20.34
botocore==1.23.34
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
countries==0.2.0
cryptography==35.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==4.0
django-bootstrap4==3.0.1
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-storages==1.12.3
entrypoints==0.3
filelock==3.3.2
fonttools==4.28.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
outcome==1.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
psycopg2==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
pygal==3.0.0
pygame==2.1.2
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
python-http-client==3.3.3
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
s3transfer==0.5.0
selenium==4.0.0
Send2Trash==1.8.0
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools-scm==6.3.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.9.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.2
wsproto==1.0.0